pyproject.toml
[project]
name = "tsg-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "pydantic>=2.11.3",
    "pytest-playwright>=0.7.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
    "scrapy>=2.12.0",
    "tld>=0.13",
    "yt-dlp>=2025.3.31",
]

[project.scripts]
scrape = "main:main"
